add decimal type to db

+++ app/actions.rb
@@ -40,7 +40,7 @@ post '/expenses/add' do
   @expense = Expense.new(
     user_id: current_user.id,
     name: params[:name],
-    value: params[:value].to_i,
+    value: params[:value].to_f,
     interval: params[:interval]
   )
 
@@ -58,7 +58,7 @@ end
 
 post '/think-twice/' do
   @expenses = Expense.where(user_id: current_user.id) if current_user
-  if params[:amount].to_i > 0
+  if params[:amount].to_f > 0
     @tt = params[:amount]
     erb :'/results'
   else
